simulation-engine/src/models/traffic_model.py
import numpy as np
import pandas as pd
from typing import Dict, List, Any
import networkx as nx
from dataclasses import dataclass
from typing import Tuple
from src.performance_optimizer import performance_optimizer

@dataclass
class TrafficMetrics:
    """Data class to hold traffic simulation metrics"""
    avg_travel_time: float
    congestion_index: float
    total_volume: float
    flow_map: Dict[str, Any]  # Simplified representation of traffic flow
    congestion_heatmap: Dict[str, Any]  # Simplified representation of congestion

class TrafficSimulationModel:
    """
    Traffic Simulation Model for calculating traffic flow, congestion levels, and travel times.
    
    This model uses graph-based road network analysis to simulate traffic patterns based on
    zoning information and population density.
    """
    
    def __init__(self):
        """Initialize the traffic simulation model"""
        self.road_network = nx.Graph()
        self.zoning_data = None
        self.population_density = 0.0
        
    def _build_road_network(self, infra_edits: List[Dict]) -> nx.Graph:
        """
        Build a graph representation of the road network from infrastructure edits.
        
        Args:
            infra_edits: List of infrastructure edits affecting roads
            
        Returns:
            NetworkX Graph representing the road network
        """
        # Create a simple graph for demonstration
        # In a real implementation, this would be built from actual road data
        G = nx.Graph()
        
        # Add nodes and edges based on infrastructure edits
        for edit in infra_edits:
            if edit.get("type") == "road":
                if edit.get("action") == "add":
                    # Add nodes and edges for new roads
                    # This is a simplified representation
                    node1 = f"node_{len(G.nodes)}"
                    node2 = f"node_{len(G.nodes)+1}"
                    G.add_edge(node1, node2, weight=1.0)
                elif edit.get("action") == "remove":
                    # Remove edges for removed roads
                    # Simplified implementation
                    pass
        
        return G
    
    def _calculate_congestion(self, road_network: nx.Graph, 
                            population_density: float) -> Dict[str, float]:
        """
        Calculate congestion levels based on road capacity and usage.
        
        Args:
            road_network: NetworkX Graph representing the road network
            population_density: Population density affecting traffic
            
        Returns:
            Dictionary with congestion metrics for each road segment
        """
        congestion_levels = {}
        
        # Simplified congestion calculation
        # In a real implementation, this would consider road capacity, traffic volume, etc.
        for edge in road_network.edges():
            # Base congestion on population density and random factors
            base_congestion = population_density * 0.001
            random_factor = np.random.uniform(0.8, 1.2)  # Random variation
            congestion_levels[f"{edge[0]}-{edge[1]}"] = base_congestion * random_factor
            
        return congestion_levels
    
    def _calculate_travel_times(self, road_network: nx.Graph, 
                              congestion_levels: Dict[str, float]) -> Dict[str, float]:
        """
        Calculate travel times based on road network and congestion levels.
        
        Args:
            road_network: NetworkX Graph representing the road network
            congestion_levels: Dictionary with congestion metrics for each road segment
            
        Returns:
            Dictionary with travel times for each road segment
        """
        travel_times = {}
        
        # Simplified travel time calculation
        for edge in road_network.edges():
            edge_key = f"{edge[0]}-{edge[1]}"
            base_time = 10.0  # Base travel time in minutes
            congestion_factor = congestion_levels.get(edge_key, 1.0)
            travel_times[edge_key] = base_time * congestion_factor
            
        return travel_times
    
    def _generate_traffic_flow_map(self, road_network: nx.Graph, 
                                 travel_times: Dict[str, float]) -> Dict[str, Any]:
        """
        Generate a simplified representation of traffic flow maps.
        
        Args:
            road_network: NetworkX Graph representing the road network
            travel_times: Dictionary with travel times for each road segment
            
        Returns:
            Dictionary representing traffic flow map
        """
        # Simplified representation of traffic flow map
        flow_map = {
            "type": "FeatureCollection",
            "features": []
        }
        
        for edge in road_network.edges():
            edge_key = f"{edge[0]}-{edge[1]}"
            flow_map["features"].append({
                "type": "Feature",
                "properties": {
                    "travel_time": travel_times.get(edge_key, 0),
                    "congestion_level": "low" if travel_times.get(edge_key, 0) < 15 else "medium" if travel_times.get(edge_key, 0) < 25 else "high"
                },
                "geometry": {
                    "type": "LineString",
                    "coordinates": [[0, 0], [1, 1]]  # Simplified coordinates
                }
            })
            
        return flow_map
    
    def _generate_congestion_heatmap(self, road_network: nx.Graph, 
                                   congestion_levels: Dict[str, float]) -> Dict[str, Any]:
        """
        Generate a simplified representation of congestion heatmaps.
        
        Args:
            road_network: NetworkX Graph representing the road network
            congestion_levels: Dictionary with congestion metrics for each road segment
            
        Returns:
            Dictionary representing congestion heatmap
        """
        # Simplified representation of congestion heatmap
        heatmap = {
            "type": "FeatureCollection",
            "features": []
        }
        
        for edge in road_network.edges():
            edge_key = f"{edge[0]}-{edge[1]}"
            congestion_value = congestion_levels.get(edge_key, 0)
            heatmap["features"].append({
                "type": "Feature",
                "properties": {
                    "congestion_index": congestion_value,
                    "level": "low" if congestion_value < 0.5 else "medium" if congestion_value < 1.0 else "high"
                },
                "geometry": {
                    "type": "Point",
                    "coordinates": [0, 0]  # Simplified coordinates
                }
            })
            
        return heatmap
    
    def run_simulation(self, zoning_plan: Dict, infra_edits: List[Dict], 
                       traffic_inputs: Dict) -> TrafficMetrics:
        """
        Run the traffic simulation model.
        
        Args:
            zoning_plan: Zoning plan information
            infra_edits: Infrastructure edits affecting roads
            traffic_inputs: Traffic-related inputs (population density, etc.)
            
        Returns:
            TrafficMetrics object with simulation results
        """
        # Store input data
        self.zoning_data = zoning_plan
        self.population_density = traffic_inputs.get("population_density", 0.0)
        
        # Build road network from infrastructure edits
        road_network = self._build_road_network(infra_edits)
        
        # Calculate congestion levels
        congestion_levels = self._calculate_congestion(road_network, self.population_density)
        
        # Calculate travel times
        travel_times = self._calculate_travel_times(road_network, congestion_levels)
        
        # Generate traffic flow map
        flow_map = self._generate_traffic_flow_map(road_network, travel_times)
        
        # Generate congestion heatmap
        congestion_heatmap = self._generate_congestion_heatmap(road_network, congestion_levels)
        
        # Calculate aggregate metrics
        avg_travel_time = np.mean(list(travel_times.values())) if travel_times else 0.0
        congestion_index = np.mean(list(congestion_levels.values())) if congestion_levels else 0.0
        total_volume = len(road_network.edges()) * self.population_density * 0.01  # Simplified volume calculation
        
        return TrafficMetrics(
            avg_travel_time=avg_travel_time,
            congestion_index=congestion_index,
            total_volume=total_volume,
            flow_map=flow_map,
            congestion_heatmap=congestion_heatmap
        )

# Example usage
if __name__ == "__main__":
    # Example of how to use the traffic simulation model
    model = TrafficSimulationModel()
    
    # Sample input data
    zoning_plan = {
        "zones": [
            {
                "id": "zone1",
                "type": "residential",
                "area": 1000.0,
                "location": {"x": 0.0, "y": 0.0},
                "properties": {}
            }
        ]
    }
    
    infra_edits = [
        {
            "type": "road",
            "action": "add",
            "specifications": {}
        }
    ]
    
    traffic_inputs = {
        "population_density": 1000.0,
        "peak_hours": ["08:00", "17:00"],
        "vehicle_ownership": 0.8
    }
    
    # Run simulation
    results = model.run_simulation(zoning_plan, infra_edits, traffic_inputs)
    
    print(f"Average Travel Time: {results.avg_travel_time:.2f} minutes")
    print(f"Congestion Index: {results.congestion_index:.2f}")
    print(f"Total Volume: {results.total_volume:.2f}")