simulation-engine/requirements.txt
fastapi==0.68.0
uvicorn[standard]==0.15.0
pandas==1.3.3
numpy==1.21.2
matplotlib==3.4.3
folium==0.12.1
pytest==6.2.5
python-multipart==0.0.5
httpx==0.23.0
PyJWT==2.4.0